+++ src/main/java/com/cheeup/converter/community/MyPostMapper.java
@@ -1,8 +1,8 @@
 package com.cheeup.converter.community;
 
 import com.cheeup.domain.community.Post;
-import com.cheeup.domain.member.Member;
 import com.cheeup.web.dto.common.Pagination;
+import com.cheeup.web.dto.community.MyPostDto;
 import com.cheeup.web.dto.community.ReadMyPostsDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -13,7 +13,7 @@ import java.util.List;
 public interface MyPostMapper {
 
     @Mapping(source = "post.id", target = "postId")
-    ReadMyPostsDto.PostResponse toPostDto(Post post, long boardId, int commentCount, ReadMyPostsDto.AuthorResponse author);
+    MyPostDto.PostResponse toPostDto(Post post, long boardId, int commentCount, MyPostDto.AuthorResponse author);
 
-    ReadMyPostsDto.Response toDto(List<ReadMyPostsDto.PostResponse> posts, Pagination pagination);
+    ReadMyPostsDto.ResponseDto toDto(List<MyPostDto.PostResponse> posts, Pagination pagination);
 }
