+++ src/main/java/com/cheeup/service/community/MyPageService.java
@@ -4,5 +4,5 @@ import com.cheeup.web.dto.community.ReadMyPostsDto;
 
 public interface MyPageService {
 
-    ReadMyPostsDto.Response getMyPosts(long id, int page, int limit);
+    ReadMyPostsDto.ResponseDto getMyPosts(long id, int page, int limit);
 }
